requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langgraph
google-api-python-client
google-auth
google-generativeai
matplotlib
numpy
pandas
seaborn
pydantic
python-dotenv
reportlab
requests
streamlit
xhtml2pdf
markdown
newsapi-python
regex 